Disable namespacing
We don't strictly need namespacing in order to have a buildset registry
serve as multiple registries.  If we don't configure it in zuul-registry
itself, and treat dockerhub images as non-namespaced, then both docker
and oci configurations can use the same buildset registry.  We would
configure docker to talk to the registry as normal and it will fetch
unqualified image names, but we would set up containers/registries.conf
to prepend urls for non-dockerhub registries, and it would not prepend
anything for dockerhub.

In this way, both configurations can use the same buildset registry.

Change-Id: I8af4ea0f5da5ad835384400108250e14a8435656